requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
python-docx
PyMuPDF
openai
python-dotenv
